refactor: share one pow_gt between counterproof guest and executor
